♻️ refactor: QueryRoot.userの処理とテストをqueries/user.rsモジュールに移動 #6
- src/queries/user.rs に User 構造体と UserQuery を移動
- UserQuery に user メソッド実装を委譲
- ユーザー関連テスト3件を queries/user.rs に移動
- lib.rs は QueryRoot の委譲とスキーマ全体テストのみ保持
- コードの責任分離を改善し、モジュール構造を整理
