+++ application/api_schema/src/queries.rs
@@ -1,0 +1,1 @@
+pub mod user;
